This is mangling the data on the second try
If I am reading this right, the whole point is to try every key in
your ssh-agent recursively before giving up, however, the data param
is getting mangled on the second try. I am trying to pass the same
parameters as they came in the original request, just using the next
key in the ssh-agent.

To test failure: before applying this patch, make your smartdc key
be the 2nd or 3rd in your ssh-agent, then try to create a new
machine.

After applying this patch, in theory it should be a no-op since the
status won't be 401 and the recursive call won't get triggered.